@@ -11,7 +11,7 @@
   "description": "React components",
   "dependencies": {
     "ramda": "^0.5.0",
-    "react": "^0.12.0"
+    "react": "^0.13.1"
   },
   "devDependencies": {
     "browserify": "~3.33.0",
